pyproject.toml
[project]
name = "globus-auth-example"
version = "0.1.0"
description = "Demonstation of an OAuth Process using Fast API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.6.9",
    "fastapi==0.135.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "python-dotenv==1.2.2",
    "uvicorn[standard]==0.41.0",
]
